requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
openai==1.30.1
python-dotenv==1.0.1
tenacity==8.3.0
structlog==24.2.0
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
geopy==2.4.1
